// frontend/src/app/api/model/adminUserDetailDto.ts
/**
 * CoreData API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RoleDto } from './roleDto';
import { AdminStudentRecordDto } from './adminStudentRecordDto';
import { StaffMiniDto } from './staffMiniDto';
import { StudentMiniDto } from './studentMiniDto';


export interface AdminUserDetailDto { 
    id?: string;
    email?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    isActive?: boolean;
    permissions?: number;
    roles?: Array<RoleDto> | null;
    student?: StudentMiniDto;
    staff?: StaffMiniDto;
    studentRecord?: AdminStudentRecordDto;
    createdAtUtc?: string;
    lastLoginAtUtc?: string | null;
}