// frontend/src/app/api/model/adminUpsertStudentRecordDto.ts
/**
 * CoreData API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AdminUpsertStudentRecordDto { 
    personalEmail?: string | null;
    homeAddress?: string | null;
    phoneNumber?: string | null;
    entryQualifications?: Array<string> | null;
    gender?: string | null;
}